finance_agent/requirements.txt
agno
groq
duckduckgo-search
google-search-results
yfinance
python-dotenv